UPDATING DATA ON PAGE LOAD
