+++ index.html
@@ -7,7 +7,7 @@
     <link rel="stylesheet" href="css/style.css" />
     <link href="https://fonts.googleapis.com/css?family=Quicksand" rel="stylesheet" />
   </head>
-  <body>
+  <body onload="getMembers()">
     <div class="Application">
       <div class="UserMenu" id="UserMenu">
         <div id="Toggleable">
